fix(sqlite): isolate private staging directory setup
