+++ src/infra/sqlite-readonly-location.ts
@@ -7,7 +7,7 @@ import {
   requireNodeSqlite,
   resolveSqliteFilesystemPath,
 } from "./node-sqlite.js";
-import { createPrivateSqliteTempDirectory } from "./sqlite-snapshot.js";
+import { createPrivateSqliteTempDirectory } from "./sqlite-private-directory.js";
 import { resolvePreferredOpenClawTmpDir } from "./tmp-openclaw-dir.js";
 
 const MAX_SNAPSHOT_ATTEMPTS = 10;
